Reinstate the bandwidth limiter server-side

+++ massa-node/src/settings.rs
@@ -97,7 +97,7 @@ pub struct BootstrapSettings {
     pub max_simultaneous_bootstraps: u32,
     pub per_ip_min_interval: MassaTime,
     pub ip_list_max_size: usize,
-    pub max_bytes_read_write: f64,
+    pub max_bytes_read_write: u128,
     /// Allocated time with which to manage the bootstrap process
     pub bootstrap_timeout: MassaTime,
 }
